+++ ElevatorAction/ElevatorAction/AbstractHitbox.hpp
@@ -9,10 +9,10 @@ public:
 	virtual float y1() const = 0;
 	virtual float x2() const = 0;
 	virtual float y2() const = 0;
-	int left();
-	int right();
-	int top();
-	int bottom();
+	float left();
+	float right();
+	float top();
+	float bottom();
 	bool collides(AbstractHitbox *h);
 };
